refactor(workflows): extract WorkflowExpressionEvaluator and DefinitionExpressionEvaluator
Move CEL evaluation out of WorkflowExecutionService.evaluateWorkflow
and DefaultStepExecutor.evaluateDefinitionExpressions into named
classes with explicit, named skip rules.

The asymmetry between the two paths is now an explicit property of
each class rather than an accidental difference between two inline
implementations:

- WorkflowExpressionEvaluator is STRICT — per-expression eval errors
  propagate. Skip rules: runtime expressions, self.* (forEach
  variables), forEach.in (kept raw for expansion), and task.inputs
  with step-output dependencies.
- DefinitionExpressionEvaluator is LENIENT — per-expression eval
  errors are swallowed; the Proxy on globalArgs surfaces a clear
  error later if the unresolved value is actually needed. Skip
  rules: runtime expressions and missing-model-dep references.

Both classes accept a CelExpressionEvaluator dependency via
constructor — a structural interface defined locally. The infrastructure
CelEvaluator implements this structurally, so the new module has zero
direct infrastructure imports (DDD ratchet stays at 26).

WorkflowExpressionEvaluator returns { workflow, expressionsEvaluated }
so callers can record the count as a tracing-span attribute (preserving
the previous inline behaviour exactly).

The dynamic `await import("../definitions/definition.ts")` is preserved
verbatim — investigating whether a top-level import is now safe is a
separate concern.

Behaviour-preserving:
- Same skip rules, same iteration order, same lenient-vs-strict split
- Same span attribute set on the workflow.evaluate span
- Verified by /tmp/swamp-verification (10 phases, OK against baseline)
  and full deno test (4999 passed)
